requirements.txt
pyshark
yaswfp
pywin32
pillow
opencv-python
numpy
matplotlib
coloredlogs